+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * build: spinoff encoder service definition
 * build: twig & twigbundle is now optional
 * feat: API Platform support
+* feat(`OpenApi`): change all 'page' parameters to accept string
 
 # 0.7.2